fix msvc warning

+++ test/setup_transfer.cpp
@@ -668,7 +668,7 @@ std::string get_python()
 	{
 		std::vector<char> buf(req_size);
 		DWORD const sz = GetEnvironmentVariable("PYTHON_INTERPRETER", buf.data(), buf.size());
-		if (sz == buf.size() - 1) return buf.data();
+		if (size_t(sz) == buf.size() - 1) return buf.data();
 	}
 #endif
 	return "python3";
